Move getUserLocale from number-format to lib
